provider: remove santity check in ECNamedProp::GetIDsFromNames
The santity check was originally added to cope with problems in
Outlook. However, the code handles NULL import very well, and it
should not return an error whenever some of the input is NULL.
It will return a warning now which will not make the entire
copy function to fail.

References: KC-345