pyko: try to open embedded messages read-write
For a certain public store at a customer, the SYSTEM user is unable
open embedded messages read-only (MAPIErrorNoAccess!), but _is_ able
to open them read-write (MAPI_MODIFY..).

While this seems like a bug in core, unfortunately I haven't been
able to reproduce the issue locally.

We do however want to open most things in python-kopano in read-
write mode at first, then try read-only mode if that fails, and
since that works in this case as well (though in an odd way), we
go for this approach for now.

References: KC-679