php: rework pointer value storing and cleanup
This reverts commit bp/8.2.0-218-g02e4c5f5 "php: add missing pointer
type conversions when calling zend_register_resource". That commit
had changed php-ext to store ptr-to-IUnknown (rather than say
ptr-to-IMAPIFolder) for the sake of the _php_free_mapi_object
function.

Even though UOBJ_REGISTER_RESOURCE converted to IUnknown first and
then to ptr-to-void, the ZEND_FETCH_RESOURCE macro lacked the
intermediate step and went straight from ptr-to-void to
ptr-to-IMAPIFolder rather than going through ptr-to-IUnknown.

So go back to storing ptr-to-IMAPIFolder in le_mapi_table and add a
bunch of destructors for each type, since that seems a bit less code
than trying to funnel ZEND_FETCH_RESOURCE do the IUnknown jump.

References: KC-673