client: avoid using unknown interface in ECExchangeImportContentsChanges
This goes in the same direction as bp/8.4.0-648-g3ff242b0 . If a
caller sets an interface different from IID_IMessage with
SetMessageInterface, then the ImportMessageChange function will not
know what to make of the pointer returned from OpenEntry.

So, the message will always be opened with IID_IMessage instead, and
the cast to the desired pointer will happen at the end of the
function only (and once in-between to see if the interface is really
supported).