libserver: devirtualize calls to non-overridden trivial return functions
Counting the number of "CALL" assembler instructions in
libkcserver.so [gcc-7].

          -O0    -O2    -Os
---------------------------
base    58702  32643  29769
1/4     58621  32613  29737
2/4     58570  32599  29715
3/4     58525  32583  29699
4/4     58466  32568  29671
devirt  58464  32178  29309
---------------------------
delta    -238   -465   -460

A command like "ecsesion->GetSessionManager()->GetCacheManager()"
used to require 5 x86_64 instructions, now it is 2. There is at least
an 18% speed improvement to be had (*if and when* you call GCM, which
does not happen often in the grand scheme of things).

References: KC-693