hl: switch object_ptr to eager deletion
As written in “The Drawbacks of Implementing Move Assignment in Terms
of Swap” on
http://scottmeyers.blogspot.de/2014/06/the-drawbacks-of-implementing-move.html

"""Regardless of which move assignment implementation is used, the
delete will eventually be performed. […] The cost therefore doesn't
vary between the implementations, but _when_ you incur that cost
does. […] it could certainly increase the maximum amount of memory
used by the application at any given time."""