pyko: fix lazy loading regression
Before the lazy loading changes, store.folder(entryid=..) would raise
an error. But because the actual mapiobj is now opened lazily, that is
is no longer the case. So add a flag to control lazy loading, and only
enable it in batch mode (that is, from folder.folders()).

While at it, make {store, folder}.item(entryid=..) also raise NotFoundError.