inetmapi: avoid placing CRLF into PS_EC_IMAP property
On mail reception, the PS_EC_IMAP property is filled with the IMAP
ENVELOPE summary representation of the mail headers. libvmime is used
to construct parts of the IMAP envelope, which includes the subject
field. If the subject is sufficiently long enough, vmime will by
default wrap it and insert CRLF to heed line length recommendations.

These CRLF sequences are in fact unwanted in PS_EC_IMAP. Have
inetmapi request that vmime not generate any line wrappings anymore.

References: KC-668