common: modernize ECTableRow to use std::string as container
Keeping three pointers to arrays (sortlen, sortkey, sortflags)
together is quite cumbersome. std::string will be used as new
length-tracked buffer for the key, which also takes care of object
copies.

Together with flags, the key is composed into a ECSortCol class.

sortflags[a] becomes arr[a].flags, sortkey[a] becomes
arr[a].key(.c_str()), and sortlen[a] is given by arr[a].key.size().