tools: print PR_ mnemonic even when proptype changes
When a property is not found, the 16-bit type portion of the 32-bit
prop id is replaced with PT_ERROR. That new prop id generally has no
associated PR_ mnemonic and so would be displayed in numeric notation
instead.

In case the entire propid is not mapped to a variable, try evaluating
just the proptag. The new output is:

(gdb) p r0.lpProps[0]
$3 = {PR_MDB_PROVIDER PT_ERROR -2147221233, …}