mapi: make GUIDs always have LE layout in memory
GUIDs in the MySQL database are stored in LE order as the DCE UUID
standards prefers it, and KC uses a byte-block comparison with memcmp
to check for equality between such a stored UUID and a right-hand
side operand. The RHS (usually a variable or pointer thereto) however
is in host order instead.

The easy way out here seems to be to just byteswap the
representations of the GUID variables in question. That way, there is
no need for byteswapping at runtime. Only downside: the textual
representation in gdb will appear swapped.

(LE)
(gdb) p/x IID_IMailUser
$3 = {Data1 = 0x2030a, Data2 = 0x0, Data3 = 0x0,
      Data4 = {0xc0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x46}}
(gdb) x/16bx &IID_IMailUser
0x7ffff44ee270 <IID_IMailUser>:   0x0a 0x03 0x02 0x00 0x00 0x00 0x00 0x00
0x7ffff44ee278 <IID_IMailUser+8>: 0xc0 0x00 0x00 0x00 0x00 0x00 0x00 0x46

(BE prospected)
(gdb) p/x IID_IMailUser
$3 = {Data1 = 0xa030200, Data2 = 0x0, Data3 = 0x0,
      Data4 = {0xc0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x46}}

References: KC-743