build: move plugins to ${libdir}
Bo Simonsen suggests: "Wouldn't it be possible to search
LD_LIBRARY_PATH for these libraries?". That is to mean that — given
"normal" invocations of the server from the system process manager
start out with no LD_LIBRARY_PATH — that the plugins be located in a
*standard location*. That way, the "plugin_path" configuration file
directive will not be needed anymore, not even for the testsuite,
since the TS already uses LD_LIBRARY_PATH for other reasons.