programs: disallow empty values for logging directives
When using, for example, log_file with no value, then ECLogger does
this:

	Unable to open logfile '' as user 'kopano'
	Not enough permissions to append logfile ''. Reverting to stderr.

That is a bit dull, so simply require that the admin specifies a
proper logfile, or "-" if he wants stderr.

Likewise, there are other options for which the empty value does not
make a lot of sense, such as log_level. If the admin wants the
default, the directive should be completely commented out.

Inspired by KC-745.