server: update PR_LOCAL_COMMIT_MAX on hard-deletes
As checked against exchange, PR_LOCAL_COMMIT_MAX should actually
change for hard-deleted messages, even though we purposefully
do not seem to do this at the moment.

It is really useful in any case to have a single property to
indicate when the (top-level) contents of a folder has last changed
(for any kind of syncing client, such as z-push).

References: KC-770