libserver: avoid creating multi-stream gzip files
The culprit is Z_FINISH. zlib.h only says: """If the flush parameter
is Z_FINISH, the remaining data is written and the gzip stream is
completed in the output."""

zlib's source code, gzwrite.c, worded it quite differently: """If
flush is Z_FINISH, then the deflate() state is reset to start a new
gzip stream."""

Completing stream 0 is one thing. Starting stream 1 is another.

References: KC-104, KC-314, KC-597