server: complete signal blocking
kcsrv_blocksigs constructed a signal mask but never used it. Fix that
up, and make sure that all threads also apply this mask, as there is
no need to interrupt the workers for signal handling.

This should aid in working around a currently not fully solved
potential race condition where reloading of the config might cause
part of the soap structure's data to become invalid and end in a
crash.

References: KC-171, KC-779