mapidump: deterministally dump mapi data
Add script to more or less deterministically dump mapi-level
data to stdout, for e.g. a store or folder.

In combination with 'diff' this makes it easy to compare mapi-level
data which should be mostly identical (think backup-restore,
comparing the PHP version of mr-accept with the one in Python,
running the PST migrator with python2 and 3..)