daemons: use empty directory as cwd
In a typical unixish distribution, daemons are started or configured with cwd=/
so that they do not needlessy hold references to mountpoints. Manpage:
daemon(7), item 11.

In case of sensitive daemons (or parts thereof) like sshd or named, these
(sub)processes are set to even run with a chroot.

Kopano services usually still want access to /etc and /usr, so a chroot is out
of the question currently. As an interim step to cwd=/, an empty subdirectory
/var/lib/kopano/empty will be used to ensure there are no relative paths (no
forward-looking ones, anyway) left anywhere.