inetmapi: generate <invalid@invalid> instead of <"@"@localhost>
The ruleset/state machine of the vmime::emailAddress parser is such
that the first character is always considered part of the local name.
(This because the local name is not allowed to be of length 0, '@'
cannot legally be used as both the first character and separator at
the same time.)

This means KC should not try to use a faulty address in the first
place, but stick to something like "invalid@invalid", which is
exactly what VMIME uses when passing in the empty string.

References: KC-689, KC-772