pyko: add rw->r fallback to utils.openentry_raw
On some systems, especially older ones it seems, it is sometimes
possible to open an item read-write but not read-only (we've seen
this before, in different contexts).

Since we want python-kopano to try both anyway, we change openentry_raw
to first try read-write, then fall back to read-only.

References: KC-820