swig: include std_string.i before kopano/typemap.i
Kopano/typemap.i requires std_string.i to be already included,
or it will cause a memory leak in the generated code (seen when
just importing inetmapi under valgrind).

So add back in the include, which was previously removed to
improve python3 support, and add a 'freearg' typemap, to override
the now included std_string.i version.