Revert "gateway: ensure that we do not swap when the query is out of bound"
The term "ulBeginMailnr <= lstFolderMailEIDs.size()" is incorrect.
ulBeginMailnr is not an index into lstFolderMailEIDs, but a UID.

We must ensure that "ulBeginMailnr <= ulMailnr" and that the upper
bound search for e starts no earlier than b, otherwise "b > e" would
hold, which then later causes the loop to never terminate and
eventually die because of memory exhaustion or illegal memory access.

Swapping is allowed per the RFC. If it were not, an
"if (ulBeginMailNr > ulMailNr) continue;" would be appropriate.

Fixes: bp/8.3.0-821-g1f71838a
References: KS-38641