server: remove dead distributed_kopano_check
Remove some very weird schema updates.

When a database is changed from schema 26 to 27 (or later), the
lock_distributed flag is added to the "settings" table. At the same
time however, if one started with a fresh v27 database (or later),
the flag would be absent. So ultimately, someone probably forgot to
update the "fresh db" code path — that makes the code testing for the
flag practically useless.