libserver: avoid hex numbers in SQL statements
In SQL, "0x8501" is treated as VARBINARY, not INTEGER. It does not
have the same behavior as in C! As for type promotion that happens
when a VARBINARY is used together with something else,  the
MySQL/MariaDB doc states "[…] the arguments are compared as
floating-point (real) numbers.". This seems to apply to
non-comparisons operations like addition and subtraction as well.

Curiously, some database installations do _not_ do this FP treatment.

typeof(0x1+0x1) is DOUBLE as documented:
	* CentOS 6 mysql-server-5.1.73-8.el6_8.x86_64
	* Debian 7 mysql-server-core-5.5_5.5.57-0+deb7u1
	* Debian 8 mysql-server-core-5.5_5.5.57-0+deb8u1
	* Ubuntu 14.04 mysql-server-core-5.5_5.5.57-0ubuntu0.14.04.1
	* openSUSE Leap 42.2 mysql-community-server-5.6.37
	* Debian 8 mariadb-server-core-10.0_10.0.32-0+deb8u1

typeof(0x1+0x1) is UNSIGNED INTEGER like in C:
	* CentOS 7 mariadb-server-5.5.52-1.el7.x86_64
	* CentOS 7 mariadb-server-5.5.56-2.el7.x86_64
	* Ubuntu 16.04 mysql-server-core-5.7_5.7.19-0ubuntu0.16.04.1
	* openSUSE Tumbleweed mariadb-10.1.2

The theory at this point is that the new behavior was implemented
around 5.7/10.1 and backported to prior generations.

Since the Kopano database seldomly (if at all?) has to deal with
integer values larger than 2^32 (let alone 9.007199e+15), there is no
immediate obvious problem manifestation, but let's fix the FP
conversions anyway.

References: KC-841