gateway: revert premature optimization for folder retrieval
The folder list retrieval function was refactored to use CONVENIENT_DEPTH
for getting the entire hierarchy with one MAPI call.
At that time, searching for a folder was also refactored to use the new
folder list retrieval function. However, that turned out to be
costly when dealing with big folder hierarchies. In theory this should
be faster because of one MAPI call and not several, however the MAPI
call it self is expensive with big folder hierarchies.

References: KC-853