gateway: added support for \Junk in IMAP LIST command
In the special folders map we store PR_IPM_<type>_ENTRYID for
every special folder where the type can be SENTMAIL, WASTEBASKET or
DRAFTS. However, there is no such constant for the Junk mail folder.
It is stored in the PR_ADDITIONAL_REN_ENTRYIDS array. Internally in
the IMAP class we use that constant for representing the Junk mail
folder, and denotes it PR_IPM_FAKEJUNK_ENTRYID.

References: KC-857