dagent: change some raw pointers to object_ptr
Raw pointers indeed need a raw AddRef, but this feels odd when there
is no directly visible matching raw Release call (hint: it is in the
caller which ties it to a object_ptr). To alleviate this cosmetic
issue, convert the three raw pinters to object_ptr too to iron it all
out.