dagent: auto-deallocate DeliveryArgs::lpChannel
lpChannel was not freed. It was not a problem, because dagent
currently only supports the "fork" process model and the worker
subprocesses just exit, but it is likely to show up in valgrind, and
in case a the "thread" model is ever reactivated. Transform it into a
unique_ptr so that it gets freed in time.

Because unique_ptrs cannot be copied, but the DeliveryArgs object is,
a slightly longer than usual constructor forms.