misc: add LIMIT clause for SELECT when only one row is evaluated
Continued earlier work on adding LIMIT clauses to SELECT statements
to make the work easier for the query optimizer.

If only one row is ever fetched from the result set, the statement is
a candidate for a LIMIT 1 addition. (Skipped for statements involving
SUM(), MAX() or MIN() and not involving GROUP BY, because such
statements will only return one row anyway.)

References: KC-5