ldapplugin: avoid triggering crash in DES_fcrypt
OpenSSL before 1.1.0~pre6, as well as variants branched therefrom,
such as libressl <= 2.6.2, crash when the salt contains a non-ASCII
character.

Because a DN's "userPassword" attribute can be edited by
authenticating with the very identity, a user can inject a custom
value with an offending character such as \xc0. The LDAP server will
not prohibit this, since the attribute is declared by RFC to be an
octet string, rather than, say, an IA5String (ASCII string).

A kopano-server configured with user_plugin=ldap and
ldap_authentication_method=password could therefore be steered into
an out-of-bounds read/crash, causing a denial of service.