ldapplugin: fix out-of-bounds in b64_encode
When the input length is a multiple of 3, b64_encode does a round too
much. Replace all invocations by base64_encode.

This is currently not exploitable, since MD5 is 16 bytes, SMD5 is 20
bytes, SHA is 20 bytes, and SSHA… is, as it turns out, completely
misconstructed, emitting only 20 instead of 24 bytes.