spooler: replace KCHL wrappers with MAPI code
The KCHL wrappers and their idiom of throwing for any non-hrSuccess exit turned
out to be an unpopular choice and they did not get much use. Understandably so,
because quite often, errors are the norm rather than an exception, such as with
NOT_FOUND cases (even vmime changed from exceptions to simpler returns).

At the same time, the goal of doing more RAII and overall less code has already
been reached with memory_ptr/object_ptr and kc_perror. Replace the few users of
KCHL wrappers with native MAPI code, which, on the grand scale, is a lot shorter
(~24 lines more in gateway+spooler, but 882 less overall, considering hl.cpp).