client: resolve cov-scan ignored return value reports
The return value correlates with the pointer value (if retval is
hrSuccess, the pointer is always valid, and if the pointer is not
nullptr, the retval must have been hrSuccess). However, to soothe
cov-scan, verify the return code instead of the pointer.

CID-123021 (#1 of 1): Unchecked return value (CHECKED_RETURN)
6. check_return: Calling HrGetOneProp without checking return
value (as is done elsewhere 314 out of 333 times).
CID-157353 (#2 of 2): Unchecked return value (CHECKED_RETURN)
CID-157367 (#1 of 1): Unchecked return value (CHECKED_RETURN)
2. check_return: Calling CopyMAPIEntryIdToSOAPEntryId without checking
return value (as is done elsewhere 60 out of 61 times). (+1 undiscovered)
CID-157375 (#1 of 1): Unchecked return value (CHECKED_RETURN)
7. check_return: Calling HrGetOneProp without checking return value
(as is done elsewhere 314 out of 333 times).

References: KC-752