treewide: mark cases of ignored return values for cov-scan
In some cases, checking the value of HrGetOneProp is a bit infeasible
(we would have to retain it), so just check the pointer and be
explicit about it with a comment.

CID-162638 (#1 of 1): Unchecked return value (CHECKED_RETURN)
1. check_return: Calling HrGetOneProp without checking return value
(as is done elsewhere 314 out of 333 times).
CID-177465 (#2 of 2): Unchecked return value (CHECKED_RETURN)
CID-177467 (#1 of 1): Unchecked return value (CHECKED_RETURN)

References: KC-752