Revert "libserver: stop relying on NO_UNSIGNED_SUBTRACTIONS"
This reverts commit f1947df9f1cdf9359b2b29de8a1187b2f20e8be4.

The expression "names.id + n" currently changes the execution plan
such that the join type changes from "eq_ref" to "ALL", which leads
to horrible performance, as all rows of "names" are rescanned for
each matching entry.

COUNT(*) of names is, for example, 6290 for KS-39199, 1877
for kopano.com, and 22 on a development system.

References: KC-841, KC-869, KS-39198, KS-39199