common, client: when attempting login do not try symmetric encrypted password
The Linux client never generated frobnicated passwords, only the Windows
portions of zarafa6client does in certain spots. Therefore, the
WSTransport::HrLogon2 function need no longer test for having received
such a password as a function argument. This happens to save an extra
RPC call when the login to the server fails.