pyko: make folder.items preload common properties
Make folder.items() preload the following common properties:

subject, last_modified, received, entryid

This makes the following code approx. 50 times faster (from about 250
items/sec to about 10000 items/sec on my test setup):

for item in folder:
    item.subject, item.entryid, item.last_modified, item.received

The main complication is that we now have to check for 'not-found'
errors and the maximum table cell length of 255 (fall back to opening
items one-by-one if this is reached).

Note also that this only works for above (item) properties, although
it can probably be generalized, and ultimately perhaps even be made
dynamic, so data is preloaded according to observed behaviour.

I don't know at this point how to be fast for properties which are
usually larger than 255 characters (or bytes?), such as bodies, without
extending the server.