libserver: add flag to allow logon without storing a session
For simple checks of username/password it is not required to actually
keep track of the session on the server and just flooding the session
table without reason only to clean those sessions up later again.
This change adds a logon flag which when set, does not store the created
session in the session table, making the logon essentially the same as
if followed by an immediate logout but without all the overhead.