inetmapi: overwrite recipients instead of appending
When converting RFC2822 mails to a MAPI object the code was appending
recipients on the MAPI object. In most cases, this is fine since conversion
takes place on clean objects. There are some edge cases where it makes
sense to reuse the object, and since all other properties are overwritten
we do the same for the recipients.

References: KC-419