libserver: add singleinstances.filename column to SQL schema
kopano-server implements a trivial server-local deduplication, which
is used, for example, when making a copy of a message. This leads to
the so-called single instance id (SIID), which is also part of the
SOAP RPC interface and cannot be changed. Its presence is not a
problem, as further deduplication can be hidden behind the attachment
storage API. (Multiple SIIDs could just point to the same
file, provided files get refcounted.)

References: KC-567