pyko: optimize Folder.path again
Optimize Folder.path again, by using PR_EC_PUBLIC_IPM_SUBTREE_ENTRYID.
In contrast to PR_IPM_PUBLIC_FOLDERS_ENTRYID, this is not converted
(and that's why it may exist), so we can match it with table data,
which is also not converted (see ECMsgStorePublic.cpp for conversion
details). This also fixes a hang when getting the public subtree parent.