server: make server_listen default empty
When kopano-server was started with an old configuration file that
also used a different port, i.e. server_tcp_port=1234, it would still
attempt to bind to port 236 because of the server_listen default.
This is unwanted.

Switch the server_listen default to the empty value, similar to how
Apache httpd does it for their listening ports directives.
kopano-server needs a config file in all cases anyway, because of
MySQL credentials.