spamd: writing the spam messages to a file in the configured spam dir
The initial version of spamd invoked sa-learn directly via sudo. This
worked fine, however it required some configuration to get running, and
it would be harder to package properly. Now, we use a different
approach, we write the messages to a specific configured spam directory.
The machine running spamassassin should run cronjobs or similar to
invoke "sa-learn --spam <configured directory>" on a regular interval.

References: KC-666