libserver: fix invalid entryid emission into SQL
The "EntryId" class has a number of user-defined conversion functions:

	%1. EntryId::operator unsigned char *() const
	%2. EntryId::operator xsd__base64Binary *()
	%3. EntryId::operator const std::string &() const

Previously, EntryId::operator unsigned char *() const was chosen for
implicit conversions involving EscapeBinary(const unsigned char *).
This obvious first-class choice is because both sides match with
regard to the basic type, and char * is directly convertible to const
char *.

When EscapeBinary was changed to take (const void *), second-class
choices would be made. The thought was that options %1 and %2 would
participate (and generate a compiler diagnostic), but alas, this was
not the case.

The cv/ref qualifiers on member functions also play a role. Because
cop.sNewEntryId is not const, option %1, because of the const
qualifier, became a third-class choice only.

Rather than entryids, the bytewise contents of a xsd__base64Binary
were sent to the SQL database, plus some uninitialized memory, since
xsd__base64Binary is usually 8/16 bytes long, while EscapeBinary was
being passed count=48.

The shortest fix is to add a void * UDC which will be a first-class
match in the resolution process.

Fixes: kopanocore-8.4.90-717-gdea32bce
References: KC-909