mapitime: rewrite login speed tests
The "lsr" login speed tests checked the time for an action sequence
that now look kind of strange. For example, "lsr+ping" was measuring
{ save state; close session; restore from state; someping; }. This is
not a sequence WebApp would do at all. Furthermore, the lsr+ping
sequence did not logically match up with the "ping" sequence { open
new session; someping; }.

"lsr+ping" is replaced by "open2" which measures { restore from
state; someping; }. The quotient open2÷open1 is expected to go up.

The kind of RPC ping has also been replaced to not issue a trivial
IECTestProtocol RPC anymore, but { OpenMsgStore; OpenRootContainer; }
instead, because the latter, too, is more likely what a MAPI client
would be doing normally. As this does a little more work than
IECTestProtocol, observed "op/s" values are expected to go down ( and
that is not a bad thing).

As such, the observed runtimes expectedly change in the downward
direction. The current values are
 - open1: 60 op/s
 - open2: 102 op/s

Which gives a Session Restorer efficiency of 70% over the initial
login (up from the now-meaningless (see above) value of 40% measured
earlier with lsr+ping).

References: KC-921