gateway: derive BinaryArray from SBinary
The only difference between SBinary and BinaryArray is the extra
member "cheap". By making BinaryArray a subclass of SBinary,

  - Comparing two BinaryArrays will re-use the pre-existing SBinary
    comparison function(s) [m4lcommon/CommonUtil.cpp]

  - Comparing an SBinary with a BinaryArray will also just use that,
    and no longer needing to invoke the BinaryArray(SBinary&) ctor
    first before BinaryArray::operator==.

  - Some constructors/assignment operators can be dropped.

It turns out that this class is not so suitable for conversion to use
std::string as an underlying type. Many places in gateway/ want to
directly access the cb and lpb fields, and in fact, even take an
address of cb (for IStream::Write).