mapitime: turn signal handler into a thread with locking
It is possible that a signal interrupts the main flow right when a
std::list is being updated, giving potentially false readings. To fix
that, the mpt_stat_dump is made into a separately-running thread, and
the list is protected with a mutex so that none works with the list
at the same time.

References: KC-921