spooler: removed always_send_utf8 from spooler config
The always_send_utf8 option causes text/* MIME parts to be transcoded
to another character set (the Content-Type header is appropriately
changed). This is fine in the realm of MIME, and e-mail clients are
able to properly read this MIME stream. However, when a MIME part is
"exported" - to a file or a even just another buffer in the same
program - a byte copy is normally performed and the information from
the Content-Type is not copied, since there is no place defined for such.
At that point, clients / viewers will interpret the bytes using the content
alone. If the content is, for example, a HTML file with a <meta> tag that
defines iso-8859-1, the used HTML widget will interpret it that way. This
will show mojibake since the byte stream was, and still is, UTF-8.

This is probably why there has always been such an odd back-and-forth in
VMIMEToMAPI::handleHTMLTextpart, sometimes favoring <meta> over Content-Type
and vice-versa. As of 2017, handleHTMLTextpart does the right thing RFC-wise,
but of course kopano should at best not even emit mails where such a problem
can arise.

always_send_utf8 is by this commit obsoleted and behavior is "no".

References: KC-901