pyko: make global store cache per-server-instance
The currently used LRU cache does not use weak references, but
it does refer to server instances, meaning the respective
socket connections aren't closed until the references are evicted
from the cache.

For kopano-rest, which uses a fresh server instance for each request,
this means that the kopano server gets to deal with 128 open connections
per worker process, which is a lot for many workers.

We could use weak references, but it seems nicer to have a cache
per server instance (which was also the original idea). This probably
means we can also remove the ugly atexit handler.