gateway: removed some more Blackberry related code
It seems that it was only blackberry needing PR_SENTMAIL_ENTRYID
on the sent mail folder. Everything else will use PR_IPM_SENTMAIL_ENTRYID.

In general, the PR_SENTMAIL_ENTRYID are used by messages in the outbox,
to specify where it is copied to, after it is sent. It does not apply
for folders.