inetmapi, spooler: introduce indexed_headers for whitelisting X headers
So e-mail systems may produce a vast amount of X headers including
Microsoft (Office 365) systems. The usual behavior was to index every
X header in a named property in the namespace PS_INTERNET_HEADERS.
However, with that vast amount of headers we get a big amount of
properties for every e-mail. An example email had totally 101 properties
with the old behavior, and 71 with just X-Spam whitelisted which were
just 4 properties.

References: KC-948