libserver: increase MAX_PROP_SIZE
The MAX_PROP_SIZE was set to 8192 bytes. This would mean whenever
a property value exceeded that value on the LoadObject call a seperate
call would be made to fetch the property. Since client and server
communicates on a usual fast network we have increased it to 32 KB,
avoiding seperate SOAP calls for bigger properties (like the WebApp
JSON config dictionary and transport headers).