client, libserver: avoiding LoadObject on GetHierarchyTable
When OpenEntry is invoked with MAPI_DEFERRED_ERRORS the properties
are not loaded immediately. By traversing a folder hierarchy we can
avoid the LoadObject of the objects in the path, by first opening
with MAPI_DEFERRED_ERRORS and then calling GetHierarchyTable. However
the client did request the properties for checking if the folder was
a search folder which caused a LoadObject. The check has been moved
to the server, where the check was anyway done, if the folder is a
search folder.