common: walk RTF by number of bytes
This changes the way the RTF data is iterated; rather than trimming
NULs and analyzing that temporary string, analyze the original string
and keep track of where the iterator is. NULs are still not copied,
but at least \binNNNN{} sections now do not skip too much.

Remove string_strip_nuls(char *) which is no longer needed because of
this.

Fixes: bp/8.2.1-916-g6d6e97c5
References: KC-969