server: fix value overlap of PROJECT_VERSION_REVISION
globalrev was used in two ways:

	1. to avoid combining a server version with a mismatching
	plugin version => so it was used in an equality test

	2. to check whether the server would attempt to use an old
	schema with a DB that is already newer by testing for
	(major,minor,micro,globalrev) with >=.

In the uint32_t globalrev, micro is only allotted 4 bits, but values
up to 99 require more than that, causing adjacent bits to get set.
8.4.90.1234 became 0x805a04d2, which is shared by 8.5.10.1234.

For (1), a string of arbitrary length can be used instead, and in
case (2), the rev need not include (major,minor,micro).