libserver: allowing users to send as a shared store again
Whenever kopanoSendAsPrivilege was used with a shared store (seen as
a non-active user), the value was simply ignored since only active users
and distlists were valid values. It has been possible to send as a
shared store with ZCP 6.40 there was the following conditions:

    strQuery = "SELECT id FROM users WHERE (object_type = "+
                stringify(USEROBJECT_TYPE_USER)+" OR object_type = "+
                stringify(USEROBJECT_TYPE_NONACTIVE)+") AND externid in (";

Therefore NONACTIVE_USER has been added as a valid objclass.

References: KC-972