pyko: allowing services to run with config defaults
If kopano.Service was used, or just kopano.Config, a config file
was required. However, all the C++ based services of KC was changed
to run without a configuration file (if possible, just relying on the
defaults), therefore the same behavior for Python based services is
desired.

References: KC-978, KC-681