installer: moved the kopano-mr-{accept|process}.py to ECtools/utils
The py.in scripts was the actual scripts which meant you would need
to run autoconf in order to make modifications. Since we get more and
more smaller tools in Python (ported from php or bash) it makes sense
to make a kopano_utils package where these can be placed.