pyko: added Item.sender setter
When generating:

    item = kopano.Server().user("user1").inbox.create_item(
        from_="user1@domain.com", to="user1@domain.com", ...
    )

It gives "Sender: invalid@invalid" in the header. That is, because
PR_SENDER_* needs to be set. That is the real "From" address and the
PR_SENT_REPRESENTING_* is the "Sender" equivalent. Now we can supply
the sender to the create_item, this should be set to the same value as
"from_". Ideally, from_ and sender should be swapped in the pyko
implementation.

References: KC-989