inetmapi: use PR_SEARCH_KEY for Message-ID generation
The Message-ID must be unique for every message (and MAPI object). The
method used for generating has produced a random Message-ID everytime
a MAPI object was converted into a RFC2822 message. Now, we uses the
PR_SEARCH_KEY property for generating the Message-ID giving the same ID
for every MAPI object.