spamd: initial support for ham learning
sa-learn has support for learning both spam and ham. Whenever a message
has been learned as spam, and it turns out to be wrong so it should be
learned as ham instead, spamassassin will forget about the spam learning.
We assume messages that we already saw in the Junk folder that is moved
to another folder will be ham message. We can probably not support the
"report_safe" option in spamassassin, therefore we will need to check for
that.

References: KC-666