client: require accurate provider type for EC{AB,MAPI}Prop generation
Even though ECGenericProp needs to deal with an arbitrary generic
provider, the subclasses do not, and therefore need not work with a
void *.

This change provides a stronger guarantee that the pointer returned
by e.g. GetABStore/GetMsgStore is, indeed, an AB/message store — by
making sure exactly that was put into lpProvider.