inetmapi: add "header_strict_rfc" to delivery options
IMToMAPI does not do any validation at all of the input, eventhough it is
required by RFC2822 to have at least "Date:" and "From:". For some clients
we may want some validation (like WebApp EML import or the gateway on
APPEND), and for some we do not. A delivery option has been added that indicate
if validation should be performed. If From or Date is missing the call
will return MAPI_E_CALL_FAILED.

References: KW-2333