pyko: Add item.reply functionality
Add a simple item.reply function, which creates a reply messages and
saves this by default in the drafts folder. A new body/subject is
generated for the new message and the source message information is
stored in a named property. When the reply is sent the soure message
information contains the reply action (reply/replyall/forward) and the
original entryid, then the correct icon index and last verb info is set
to show that a repl has been made on the original item and on what time.