libserver: do not emit empty http(s) URIs in SOAP when undefined
When a server object in LDAP declares no SSL port at all, the
serverdetails_t structure keeps ulSslPort=0 around. When requesting
GetSslPath(), a zero-length std::string is returned and used to
populate the SOAP response. It is better to not populate the XML tag
at all to indicate that it is unset.

When running kopano-archiver the log messages change from

[info   ] Check server: "ms1.lh" ssl="" flag=00000000
[crit   ] Cannot access : No such file or directory
[error  ] gsoap connect: ()
[error  ] virtual HRESULT M4LMsgServiceAdmin::ConfigureMsgService(…):
MSGServiceEntry failed: network error (80040115)

to a more meaningful

[info   ] Check server: "ms1.lh" ssl="<UNKNOWN>" flag=00000000
[error  ] No SSL or File path found for server: "ms1.lh", please fix
your configuration.

References: KC-1035