libserver: handle duplicate named property requests
Asking for a named property more than once ought to yield (0,0,0) on
not found, e.g. (2,2,2) when found, and (2,2,2) when
flags=MAPI_CREATE:

	p = MAPINAMEID(PSETID_Archive, MNID_STRING, u'store-entryids')
	mapiobj.GetIDsFromNames([p, p, p], flags)

Before b3bf36fb, each element was handled individually with an SQL
query. If an INSERT happened because of flags=MAPI_CREATE, all
subsequent SELECTs would see it. b3bf36fb broke this two-fold: it
would return (2,0,0) when found, and (2,3,4) with MAPI_CREATE.

Fixes: kopanocore-8.4.0-freeze~278^2
References: KC-1107, KC-1108