spooler: introduce log_raw_message = error
It is quite hard to debug whenever mr-process or mr-accept fails since
the original mail is deleted. Instead of activating log_raw_message globally,
we made it possible to supply "error" as value, and the raw message will *only*
be logged if an error happens during automatic processing (mr-accept/mr-process).

References: KC-1105