dagent: some support for RFC6531 (SMTPUTF8)
Postfix 3 supports the SMTPUTF8 extension and is used if the email address
of either sender or receiver contains non-ASCII characters. If not explicitly
disabled in Postfix some emails cannot be delivered to kopano-dagent since
support was missing. Now, we announce SMTPUTF8 support since we are already
able to cope with addresses having UTF-8 characters.

References: KC-833