libserver: stop querying information_schema database
It has been observed that SELECTing on information_schema is not
always possible due to lack of permissions for the DB user. Because
the "names" table never gets any DELETE statements in normal
operation, MAX(id) correlates with AUTO_INCREMENT and therefore
provides a good alternative heuristic for whether or not K-1219 has
occurred.

(Only the new kopano-dbadm utility will issue DELETEs, and if you use
it, np-defrag should be run for good measure too, and it so happens
to align MAX(id) back with AUTO_INCREMENT again.)